# docs/api/openapi-v1-phase2.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Alter Earth API",
        "description": "A community platform for ecological conservation, sustainability advocacy, and using AI for environmental good",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/comments": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "Create Comment",
                "description": "Create a new comment on a post or reply to an existing comment.\n\n- If parent_id is null, creates a top-level comment\n- If parent_id is set, creates a reply to that comment",
                "operationId": "create_comment_api_v1_comments_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comments/post/{post_id}": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Get Post Comments",
                "description": "Get all comments for a post as a threaded tree structure.\n\nReturns top-level comments with nested replies.",
                "operationId": "get_post_comments_api_v1_comments_post__post_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentTreeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{comment_id}": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Get Comment",
                "description": "Get a single comment by ID.",
                "operationId": "get_comment_api_v1_comments__comment_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "comments"
                ],
                "summary": "Update Comment",
                "description": "Update a comment's content.\n\nOnly the comment author can update it.",
                "operationId": "update_comment_api_v1_comments__comment_id__put",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "comments"
                ],
                "summary": "Delete Comment",
                "description": "Soft delete a comment.\n\nOnly the comment author can delete it.\nPreserves thread structure by marking as deleted instead of removing.",
                "operationId": "delete_comment_api_v1_comments__comment_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/votes/comments/{comment_id}": {
            "post": {
                "tags": [
                    "votes"
                ],
                "summary": "Vote On Comment",
                "description": "Vote on a comment.\n\n- vote_value: 1 for upvote, -1 for downvote\n- If user already voted, updates the vote\n- Cannot vote on own comments\n- Vote weight is based on user's karma",
                "operationId": "vote_on_comment_api_v1_votes_comments__comment_id__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentAuthor": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "nickname": {
                        "type": "string",
                        "title": "Nickname"
                    },
                    "user_type": {
                        "type": "string",
                        "title": "User Type"
                    },
                    "is_expert": {
                        "type": "boolean",
                        "title": "Is Expert",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "nickname",
                    "user_type"
                ],
                "title": "CommentAuthor",
                "description": "Simplified user info for comment author."
            },
            "CommentCreate": {
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 1,
                        "title": "Content"
                    },
                    "parent_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parent Id",
                        "description": "Parent comment ID for replies (null for top-level)"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "CommentCreate",
                "description": "Schema for creating a comment.",
                "examples": [
                    {
                        "content": "This is a great article! Thanks for sharing."
                    },
                    {
                        "content": "I agree with your point about ocean conservation.",
                        "parent_id": "550e8400-e29b-41d4-a716-446655440000"
                    }
                ]
            },
            "CommentResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "User Id"
                    },
                    "post_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Post Id"
                    },
                    "parent_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parent Id"
                    },
                    "upvotes": {
                        "type": "integer",
                        "title": "Upvotes"
                    },
                    "downvotes": {
                        "type": "integer",
                        "title": "Downvotes"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score"
                    },
                    "is_deleted": {
                        "type": "boolean",
                        "title": "Is Deleted"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "author": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/CommentAuthor"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "user_vote": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Vote"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "content",
                    "user_id",
                    "post_id",
                    "parent_id",
                    "upvotes",
                    "downvotes",
                    "score",
                    "is_deleted",
                    "created_at",
                    "updated_at"
                ],
                "title": "CommentResponse",
                "description": "Schema for comment in responses (includes author info)."
            },
            "CommentTreeResponse": {
                "properties": {
                    "post_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Post Id"
                    },
                    "total_comments": {
                        "type": "integer",
                        "title": "Total Comments"
                    },
                    "top_level_count": {
                        "type": "integer",
                        "title": "Top Level Count"
                    },
                    "comments": {
                        "items": {
                            "$ref": "#/components/schemas/CommentWithReplies"
                        },
                        "type": "array",
                        "title": "Comments"
                    }
                },
                "type": "object",
                "required": [
                    "post_id",
                    "total_comments",
                    "top_level_count",
                    "comments"
                ],
                "title": "CommentTreeResponse",
                "description": "Schema for complete comment tree on a post.",
                "example": {
                    "comments": [],
                    "post_id": "550e8400-e29b-41d4-a716-446655440000",
                    "top_level_count": 5,
                    "total_comments": 15
                }
            },
            "CommentUpdate": {
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 1,
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "CommentUpdate",
                "description": "Schema for updating a comment.",
                "example": {
                    "content": "Updated comment content..."
                }
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VoteCreate": {
                "properties": {
                    "vote_value": {
                        "type": "integer",
                        "enum": [
                            1,
                            -1
                        ],
                        "title": "Vote Value",
                        "description": "1 for upvote, -1 for downvote"
                    }
                },
                "type": "object",
                "required": [
                    "vote_value"
                ],
                "title": "VoteCreate",
                "description": "Schema for creating or updating a vote.",
                "examples": [
                    {
                        "vote_value": 1
                    },
                    {
                        "vote_value": -1
                    }
                ]
            },
            "VoteResponse": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "User Id"
                    },
                    "post_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Post Id"
                    },
                    "comment_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Comment Id"
                    },
                    "vote_value": {
                        "type": "integer",
                        "title": "Vote Value"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "vote_value",
                    "created_at",
                    "updated_at"
                ],
                "title": "VoteResponse",
                "description": "Schema for vote in responses."
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}